requirements.txt
flask==3.0.0
twilio==8.12.0
gspread==5.12.4
google-auth==2.25.2
gunicorn==21.2.0